+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 - **Breaking**: generic: register_menu系のコールバックをFnに変更
 - **Breaking**: aviutl2-macros: `generic::menus`の引数から`&mut EditSection`を削除
 - generic: `get_host_app_window`、`get_host_app_window_raw`を追加
+- eframe: 終了処理を改善
 
 ### デモプラグイン
 
-（なし）
+- metronome、local-alias、scripts-search: ヘッダーを折りたためるように
 
 ## [0.15.1](https://github.com/sevenc-nanashi/aviutl2-rs/releases/tag/0.15.1)